src/backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2.20
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
idna==3.10
pillow==10.4.0
pip==24.3.1
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.3
urllib3==2.2.3
django-filter==24.3
daphne==4.1.2
channels==4.2.2